backend/api-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
mysql-connector-python==8.2.0
bcrypt==4.1.2
pyjwt==2.8.0
cryptography==41.0.7
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6